requirements.txt
asgiref==3.7.2
boto3==1.34.65
botocore==1.34.65
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2
django-ses
django-jazzmin==2.6.0
django-storages==1.14.2
djangorestframework==3.14.0
django-cors-headers
django-currentuser==0.7.0
idna==3.6
jmespath==1.0.1
mysqlclient==2.2.4
pillow==10.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3
django-ckeditor
django-autocomplete-light
jaraco.text
python-dotenv
elasticsearch>=7.0.0,<8.0.0
django-haystack==3.3.0
djangorestframework-simplejwt==5.3.1